Cleaning up compiler warnings in order to sort out issues.
Also trying to support more compilers on Windows in case it is a
Visual Studio specific issue that causes irBZ determination to
fail when the same code works under linux.
One should now be able to create the python module using the MinGW
g++ compiler under Windows -- this has necessitates moving from
sub-packaged pybind11 to system-wide-installed pybind11.

Other "improvements" are starting work on a brute-force irBZ
determination algorithm.

+++ src/grid4.h
@@ -296,7 +296,7 @@ public:
     slong tmp;
     for (int i=0; i<4; i++){
       tmp = (slong)( round( (x[i] - this->zero[i])/this->step[i] ) );
-      if (tmp>=0 && tmp<this->size(i)){
+      if (tmp>=0 && static_cast<size_t>(tmp)<this->size(i)){
         if (approx_scalar(this->step[i]*tmp + this->zero[i],x[i]))
           out += 1<<i; // exact match
       } else {
